Update task.md
language checked. Comments:
I have a suspicion that item 3 in the canvas generating algorithm should stay "horizontally" instead of "vertically". No?
The Project Example gif has to be edited: 
"Do you want to use a pre-defined pattern or a custom one?"
"Please input 'yes' for a pre-defined pattern or 'no' for a custom one."
"Please input 'yes' or 'no'."
"Please choose a pattern." (2 times)
""Please choose the generator."
"Please input the width of the resulting picture."
"Please input the height of the resulting picture."