Update task.md
Language checked. Comments:
The gif text has to be edited:
"Two people play the game: one chooses a word (a sequence of letters), and the other one tries to guess it. In this version, the computer chooses the word: a sequence of 4 letters (for example, ACEB). The user has several attempts to guess it (the max number is 3). "

"For example, with ACEB as the hidden word, the BCDF guess will give 1 full match (C) and 1 partial match (B)."
"Please input your guess." (5 times) 
"It should be of length 4, and each symbol should be from the alphabet: ABCDEFGH." (4 times)
"Your guess has 4 full matches and 0 partial matches."