Update task.md (#66)
* Update task.md

language checked. The text in the Project Example gif should be slightly edited: 
"At each attempt, you should enter one letter; if it is present in the hidden word, all matches will be displayed."
"For example, if the word CAT was chosen by the computer, _ _ _ will be displayed first, since the word has 3 letters."
"If you enter the letter A, you will see _ A _, and so on."
"Good luck with the game!"
"My word is: _ _ _ _."
"Please input your guess!" (6 times)
"Congratulations! You guessed it!"

* Fix typos

* Add chat info

Co-authored-by: anastasia.birillo <anastasia.i.birillo@gmail.com>